+++ geo_py_utils/census_open_data/census.py
@@ -3,17 +3,16 @@
 """ Functions to download and cache polygon boundaries for select census/canada post regions """
 
 import geopandas as gpd
-from os.path import join
 import logging
 import numpy as np
+from os.path import join
 
-
-from ben_py_utils.misc.cache import Cache_wrapper
-
-from geo_py_utils.etl.download_zip import download_zip_shp
-from geo_py_utils.misc.constants import DATA_DIR
 from geo_py_utils.census_open_data.open_data import download_qc_city_neighborhoods
+from geo_py_utils.etl.download_zip import download_zip_shp
 from geo_py_utils.geo_general.crs import get_crs_str
+from geo_py_utils.misc.cache import Cache_wrapper
+from geo_py_utils.misc.constants import DATA_DIR
+
 
 logger = logging.getLogger(__file__)
 
@@ -213,12 +212,19 @@ def download_fsas(year : int = 2016,
             zip_download_url = "https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/files-fichiers/2016/lfsa000b16a_e.zip"  \
             if use_cartographic \
             else "https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/files-fichiers/2016/lfsa000a16a_e.zip"
+        elif year == 2021:
+            zip_download_url = "https://www12.statcan.gc.ca/census-recensement/2021/geo/sip-pis/boundary-limites/files-fichiers/lfsa000b21a_e.zip"  \
+            if use_cartographic \
+            else "https://www12.statcan.gc.ca/census-recensement/2021/geo/sip-pis/boundary-limites/files-fichiers/lfsa000a21a_e.zip"
         else:
-            raise ValueError(f"Fatal error - inputed {year}, but only 2016 implemented")
+            raise ValueError(f"Fatal error - inputed {year}, but only 2016 and 2021 implemented")
 
-        shp_fsa_all = download_zip_shp(zip_download_url, data_download_path)
+        # Download zip + unzip  look for shp file    
+        shp_fsa_all = download_zip_shp(zip_download_url, data_download_path, extension='.shp')
+   
+        # Filter by province
         shp_fsa_all_filtered = shp_fsa_all.loc[shp_fsa_all.PRUID.astype('str') == str(pr_code), ]
-    
+            
         logger.info(f"There are {shp_fsa_all_filtered.shape[0]} FSAs in {pr_code} for the 2016 census")
 
         # Transform
@@ -261,28 +267,7 @@ def download_qc_city_fsa_2016(buffer_degrees=0.1, **fsa_fun_kwargs) -> gpd.GeoDa
 
     return shp_2016_fsas_qc_city
 
-
-"""
-@Cache_wrapper(path_cache=join(DATA_DIR,"qc_city_fsaldu_2016_df.parquet"))
-def download_qc_city_fsaldu_2016_df(**kwargs)-> pd.DataFrame:
-
-    # Select only the fsaldu within the fsa within the city
-    shp_2016_fsas_qc_city = download_qc_city_fsa_2016(**kwargs)
-
-    # Get the pccf file
-    df_pccf = get_df_pccf()
-
-    # Extract the FSA
-    df_pccf['FSA'] = df_pccf.POSTCODE.str[:3]
-
-    # Join to filter by province (PCCF has all Canadian PCs)
-    df_pccf_qc_city = pd.merge(df_pccf,
-            shp_2016_fsas_qc_city[['CFSAUID']]   ,
-            left_on = 'FSA' ,
-            right_on =  'CFSAUID')
-
-    return df_pccf_qc_city
-"""
+ 
 
 
  
@@ -342,5 +327,6 @@ def download_prov_boundary(year: int = 2021,
 
 if __name__ == "__main__":
      #df = download_qc_city_fsa_2016() 
-     df_ca = download_ca_cmas(pr_code=24, new_crs = 4269) 
+     #df_ca = download_ca_cmas(pr_code=24, new_crs = 4269) 
+     download_fsas(2021,  24, use_cartographic=False)