+++ tests/census_open_data/test_census_open_data.py
@@ -30,7 +30,7 @@ def test_admRegs_dissolved(clean_slate=False):
     qc_admReg_dissolved_extracter = DownloadQcDissolvedAdmReg()
     shp_admReg = qc_admReg_dissolved_extracter.get_qc_administrative_boundaries()
 
-    assert shp_admReg.shape[0] == 17
+    assert shp_admReg.shape[0] == 15
 
     assert np.all(np.isin([DownloadQcDissolvedAdmReg.ADM_REG_DISSOLVE_ID_COL, 'DOMAINE_OPUS'], shp_admReg.columns))
 
@@ -74,8 +74,9 @@ def test_qa_adm_boundaries():
     # For some reason the number of features is different bepending on the precision - 1/20 or 1/100
     dict_check_num_records = {
             DownloadQcAdmBoundaries.QC_PROV_ADM_BOUND_METRO: 2,
-            DownloadQcAdmBoundaries.QC_PROV_ADM_BOUND_ADM_REG: 106,
-            DownloadQcAdmBoundaries.QC_PROV_ADM_BOUND_MUNI: 1347
+            DownloadQcAdmBoundaries.QC_PROV_ADM_BOUND_MRC: 106,
+            DownloadQcAdmBoundaries.QC_PROV_ADM_BOUND_MUNI: 1347,
+            DownloadQcAdmBoundaries.QC_PROV_ADM_BOUND_ARROND: 41
     }
     
     dict_results_num_records = {